requirements.txt
requests==2.32.3
PyYAML==6.0.2
plexapi==4.15.8
Pillow==10.4.0
tmdbsimple==2.9.1
fanarttv==1.1.0
APScheduler==3.10.4
loguru==0.7.2
